1. Fix Issue selectCells selects cell in another AOI. selectCells now is area specific to fix an issue that clicking on a cell would set it to monitored in another area with the same path.row. cache.py:get_cell(path,row) now also takes area_name parameter. main.py:SelectCellHandler() also now takes area_name parameter and uses it in cal to get_cell. Also added /<area_name> to route for /selectcell/<area_name><celldata>
view-maps.js has a new global variable area_url. This is used here in
cellSelected() to build a url. Also used by base-maps.js

2. Dragger
base-maps.js  has changes for the overlay dragger. It starts at 10% left
when a single map (view-area) bout at 90% for an twin map (view-obs)
Started to put a gripper element onto the top of the browser but its not
visible yet.