More changes to support adding and editing border and displaying either boundary_hull (as a drawing - yellow layer) or boundary_geojson - user provided geojson as a data layer (green with black fill). Data layer is the more powerfull approach.
Change Map-view is now using the new PATCH handler so the old
UpdateAreaViewHandler and route is deleted.

There is still much to do to fix the UX for editing and updating areas -
particularly the border editor which barely works as a UX.
Also areas without a border (just an area_location marker ) still fail
to generate observations.
Also editing the border does not update the LandsatCells correctly -
delete the old ones.

Major upgrade to the models.AreaOfInterest. It now stores both the
user's geometry and a convex hull - used for Landsat overlap
calcualtions.

For the User class, it is no longer necessary to count the user's areas
as this is now derived from a query on User.

common-maps.js: Added a file that is shared b/w new-area and the  
patch_area is moved to this file.

base-area.js: is upgraded to support some of the new geojson capability.
new-area.js supports an edit-boundary mode which is for editing an
existing area - no tabs are displayed.



Also