area-descriptions.html: change follow button to watch button. updated color to btn-lt-blue. added eye icons to Watch buttons.
area.html:
Added padlock icon for private areas and open icone for public areas.
change follow button to watch button. updated color to btn-lt-blue.
Change 'Follow Area' to 'Watch Area' and 'Unfolow' to 'Stop Watching'.
This updates the screen text but the URL's and methods are still follow/unfollow.

Changes to be committed:
	modified:   templates/allareas.html
	modified:   templates/areas.html
	modified:   templates/includes/area-description.html